packages/core: group api implementations by the name of the declarations
