+++ packages/core/src/api/apis/implementations/AlertApi/index.ts
@@ -13,27 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ErrorApi, ErrorContext, AlertApi } from '../../../';
 
-/**
- * Decorates an ErrorApi by also forwarding error messages
- * to the alertApi with an 'error' severity.
- */
-export class ErrorAlerter implements ErrorApi {
-  constructor(
-    private readonly alertApi: AlertApi,
-    private readonly errorApi: ErrorApi,
-  ) {}
-
-  post(error: Error, context?: ErrorContext) {
-    if (!context?.hidden) {
-      this.alertApi.post({ message: error.message, severity: 'error' });
-    }
-
-    return this.errorApi.post(error, context);
-  }
-
-  error$() {
-    return this.errorApi.error$();
-  }
-}
+export { AlertApiForwarder } from './AlertApiForwarder';
