+++ packages/core/src/api/apis/implementations/ErrorApi/ErrorApiForwarder.ts
@@ -13,10 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ErrorApi, ErrorContext } from '../../../';
-import { PublishSubject } from './lib';
-import { Observable } from '../../types';
+import { ErrorApi, ErrorContext } from '../../../..';
+import { PublishSubject } from '../lib';
+import { Observable } from '../../../types';
 
+/**
+ * Base implementation for the ErrorApi that simply forwards errors to consumers.
+ */
 export class ErrorApiForwarder implements ErrorApi {
   private readonly subject = new PublishSubject<{
     error: Error;
